Operações básicas do RepositoryMySQL

+++ src/utils/Coordinate.h
@@ -5,41 +5,41 @@ class Coordinate {
 public:
     Coordinate(){ }
 
-    Coordinate(double x, double y, double z){
+    Coordinate(float x, float y, float z){
         this->x = x;
         this->y = y;
         this->z = z;
     }
     virtual ~Coordinate(){ }
 
-    double getX() const {
+    float getX() const {
         return x;
     }
 
-    void setX(double x) {
+    void setX(float x) {
         this->x = x;
     }
 
-    double getY() const {
+    float getY() const {
         return y;
     }
 
-    void setY(double y) {
+    void setY(float y) {
         this->y = y;
     }
 
-    double getZ() const {
+    float getZ() const {
         return z;
     }
 
-    void setZ(double z) {
+    void setZ(float z) {
         this->z = z;
     }
 
 private:
-    double x;
-    double y;
-    double z;
+    float x;
+    float y;
+    float z;
 };
 
 #endif
